Crate fedimint_api

Re-exports

pub use module::ServerModulePlugin;

pub use tiered::Tiered;

pub use crate::core::server;

pub use tiered_multi::*;

Modules

backup

bitcoin_rpc

cancellable

config

core

Fedimint Core API (common) module interface

db

encoding

This module defines a binary encoding interface which is more suitable for consensus critical encoding thant e.g. bincode. Over time all structs that need to be encoded to binary will be migrated to this interface.

macros

module

net

task

tiered

tiered_multi

Macros

__api_endpoint

Example

_dyn_newtype_define_inner

_dyn_newtype_impl_deref_mut

dyn_newtype_define

Define “dyn newtype” (a newtype over dyn Trait)

dyn_newtype_impl_dyn_clone_passhthrough

Implement Clone on a “dyn newtype”

module_dyn_newtype_impl_encode_decode

Implement Encodable and Decodable for a “module dyn newtype”

module_plugin_trait_define

Define a “plugin” trait

plugin_types_trait_impl

Implements the Plugin* traits for all associated types of a FederationServerPlugin.

serde_module_encoding_wrapper

Creates a struct that can be used to make our module-decodable structs interact with serde-based APIs (HBBFT, jsonrpsee). It creates a wrapper that holds the data as serialized bytes internally.

Structs

Amount

Represents an amount of BTC inside the system. The base denomination is milli satoshi for now, this is also why the amount type from rust-bitcoin isn’t used instead.

Feerate

OutPoint

PeerId

TransactionId

A transaction id for peg-ins, peg-outs and reissuances

Enums

ParseAmountError

Traits

BitcoinHash

Trait which applies to hashes of all types.

NumPeers

for consensus-related calculations given the number of peers