Module fedimint_api::core

Fedimint Core API (common) module interface

Fedimint supports externally implemented modules.

This (Rust) module defines common interoperability types and functionality that is used on both client and sever side.

Modules

client

encode

server

Fedimint Core Server module interface

Structs

ConsensusItem

An owned, immutable output of a Transaction before it was finalized

Input

An owned, immutable input to a Transaction

KeyPair

Opaque data structure that holds a keypair consisting of a secret and a public key.

Output

An owned, immutable output of a Transaction

OutputOutcome

An owned, immutable output of a Transaction before it was finalized

PendingOutput

An owned, immutable output of a Transaction before it was finalized

Signature

SpendableOutput

An owned, immutable output of a Transaction after it was finalized (so it’s spendable)

Transaction

Transaction that was already signed

Enums

FinalizationError

Traits

ModuleConsensusItem

ModuleDecode

ModuleInput

Something that can be an Input in a Transaction

ModuleOutput

Something that can be an Output in a Transaction

ModuleOutputOutcome

ModulePendingOutput

A pending output - tracked and persisted by the client

ModuleSpendableOutput

A spendable output - tracked and persisted by the client

PluginConsensusItem

PluginDecode

Module Decoder trait

PluginInput

PluginOutput

PluginOutputOutcome

PluginPendingOutput

PluginSpendableOutput

Type Definitions

ModuleKey

A module key identifing a module