Trait fedimint_api::core::client::ClientModulePlugin

pub trait ClientModulePlugin: Debug {
    type Decoder: PluginDecode;
    type Module: ServerModulePlugin;

    const MODULE_KEY: ModuleKey;

    fn input_amount(
        &self,
        input: &<Self::Module as ServerModulePlugin>::Input
    ) -> TransactionItemAmount;
    fn output_amount(
        &self,
        output: &<Self::Module as ServerModulePlugin>::Output
    ) -> TransactionItemAmount;
}

Required Associated Types

type Decoder: PluginDecode

type Module: ServerModulePlugin

Required Associated Constants

const MODULE_KEY: ModuleKey

Required Methods

fn input_amount(
&self,
input: &<Self::Module as ServerModulePlugin>::Input
) -> TransactionItemAmount

Returns the amount represented by the input and the fee its processing requires

fn output_amount(
&self,
output: &<Self::Module as ServerModulePlugin>::Output
) -> TransactionItemAmount

Returns the amount represented by the output and the fee its processing requires