Module fedimint_api::core::server

Fedimint Core Server module interface

Fedimint supports externally implemented modules.

This (Rust) module defines common interoperability types and functionality that are only used on the server side.

Structs

ApiEndpoint

Api Endpoint exposed by a server side module

ApiHandler

ApiEndpoint handler exposed by the server side module

InputMeta

ServerModule

VerificationCache

Enums

Error

Traits

IServerModule

Backend side module interface

InitHandle

An interface exposed to Fedimint modules

ModuleApiHandler

ModuleVerificationCache

PluginVerificationCache

RpcHandlerCtx

An interface rpc handlers can use to query to access current running context

ServerModulePlugin