Trait fedimint_api::core::ModuleSpendableOutput

pub trait ModuleSpendableOutput: DynEncodable {
    fn as_any(&self) -> &(dyn Any + '_);
    fn module_key(&self) -> ModuleKey;
    fn amount(&self) -> Amount;
    fn clone(&self) -> SpendableOutput;
    fn key(&self) -> String;
}

A spendable output - tracked and persisted by the client

Created by generating transaction Output, spendable by converting to Input.

Required Methods

fn as_any(&self) -> &(dyn Any + '_)

fn module_key(&self) -> ModuleKey

Module key

fn amount(&self) -> Amount

fn clone(&self) -> SpendableOutput

fn key(&self) -> String

Prepare Input spending thish output in a transaction, and a key used to sign the Transaction