Trait fedimint_api::core::PluginDecode

pub trait PluginDecode: Debug {
    fn clone_decoder() -> Decoder;
    fn decode_input(r: &mut dyn Read) -> Result<Input, DecodeError>;
    fn decode_output(r: &mut dyn Read) -> Result<Output, DecodeError>;
    fn decode_output_outcome(
        r: &mut dyn Read
    ) -> Result<OutputOutcome, DecodeError>;
    fn decode_consensus_item(
        r: &mut dyn Read
    ) -> Result<ConsensusItem, DecodeError>;
}

Module Decoder trait

Static-polymorphism version of ModuleDecode

All methods are static, as the decoding code is supposed to be instance-independent, at least until we start to support modules with overriden ModuleKeys

Required Methods

fn clone_decoder() -> Decoder

fn decode_input(r: &mut dyn Read) -> Result<Input, DecodeError>

Decode Input compatible with this module, after the module key prefix was already decoded

fn decode_output(r: &mut dyn Read) -> Result<Output, DecodeError>

Decode Output compatible with this module, after the module key prefix was already decoded

fn decode_output_outcome(r: &mut dyn Read) -> Result<OutputOutcome, DecodeError>

Decode OutputOutcome compatible with this module, after the module key prefix was already decoded

fn decode_consensus_item(r: &mut dyn Read) -> Result<ConsensusItem, DecodeError>

Decode ConsensusItem compatible with this module, after the module key prefix was already decoded