Trait fedimint_api::db::IDatabaseTransaction

pub trait IDatabaseTransaction<'a>: 'a {
    fn raw_insert_bytes(
        &mut self,
        key: &[u8],
        value: Vec<u8>
    ) -> Result<Option<Vec<u8>>>;
    fn raw_get_bytes(&self, key: &[u8]) -> Result<Option<Vec<u8>>>;
    fn raw_remove_entry(&mut self, key: &[u8]) -> Result<Option<Vec<u8>>>;
    fn raw_find_by_prefix(&self, key_prefix: &[u8]) -> PrefixIter<'_>;
    fn commit_tx(self: Box<Self>) -> Result<()>;
    fn rollback_tx_to_savepoint(&mut self);
    fn set_tx_savepoint(&mut self);
}

Fedimint requires that the database implementation implement Snapshot Isolation. Snapshot Isolation is a database isolation level that guarantees consistent reads from the time that the snapshot was created (at transaction creation time). Transactions with Snapshot Isolation level will only commit if there has been no write to the modified keys since the snapshot (i.e. write-write conflicts are prevented).

Specifically, Fedimint expects the database implementation to prevent the following anamolies:

Non-Readable Write: TX1 writes (K1, V1) at time t but cannot read (K1, V1) at time (t + i)

Dirty Read: TX1 is able to read TX2’s uncommitted writes.

Non-Repeatable Read: TX1 reads (K1, V1) at time t and retrieves (K1, V2) at time (t + i) where V1 != V2.

Phantom Record: TX1 retrieves X number of records for a prefix at time t and retrieves Y number of records for the same prefix at time (t + i).

Lost Writes: TX1 writes (K1, V1) at the same time as TX2 writes (K1, V2). V2 overwrites V1 as the value for K1 (write-write conflict).

Table for anamolies that are prevented in each DB implementation: | Non-Readable Write | Dirty Read | Non-Repeatable Read | Phantom Record | Lost Writes MemoryDB | X | X | X | X | SledDB | X | X | | | RocksDB | X | X | X | X | X

Required Methods

fn raw_insert_bytes(
&mut self,
key: &[u8],
value: Vec<u8>
) -> Result<Option<Vec<u8>>>

fn raw_get_bytes(&self, key: &[u8]) -> Result<Option<Vec<u8>>>

fn raw_remove_entry(&mut self, key: &[u8]) -> Result<Option<Vec<u8>>>

fn raw_find_by_prefix(&self, key_prefix: &[u8]) -> PrefixIter<'_>

fn commit_tx(self: Box<Self>) -> Result<()>

fn rollback_tx_to_savepoint(&mut self)

fn set_tx_savepoint(&mut self)

Create a savepoint during the transaction that can be rolled back to using rollback_tx_to_savepoint. Rolling back to the savepoint will atomically remove the writes that were applied since the savepoint was created.

Warning: Avoid using this in fedimint client code as not all database transaction implementations will support setting a savepoint during a transaction.