Module fedimint_api::encoding

This module defines a binary encoding interface which is more suitable for consensus critical encoding thant e.g. bincode. Over time all structs that need to be encoded to binary will be migrated to this interface.

Modules

btc 🔒

secp256k1 🔒

tbs 🔒

Macros

impl_encode_decode_num 🔒

impl_encode_decode_tuple 🔒

Structs

CountWrite 🔒

A wrapper counting bytes written

DecodeError

SerdeEncodable

Wrappers for T that are De-Serializable, while we need them in Encodable contex

Traits

Decodable

Data which can be encoded in a consensus-consistent way

DynEncodable

Object-safe trait for things that can encode themselves

Encodable

Data which can be encoded in a consensus-consistent way

Derive Macros

Decodable

Encodable

UnzipConsensus