Trait fedimint_api::module::FederationModule

pub trait FederationModule: Sized {
    type Error;
    type TxInput: Send + Sync;
    type TxOutput;
    type TxOutputOutcome;
    type ConsensusItem;
    type VerificationCache;

    fn await_consensus_proposal<'a, 'async_trait>(
        &'a self,
        rng: impl 'async_trait + RngCore + CryptoRng + 'a
    ) -> Pin<Box<dyn Future<Output = ()> + 'async_trait>>
    where
        'a: 'async_trait,
        Self: 'async_trait;
    fn consensus_proposal<'a, 'async_trait>(
        &'a self,
        rng: impl 'async_trait + RngCore + CryptoRng + 'a
    ) -> Pin<Box<dyn Future<Output = Vec<Self::ConsensusItem>> + 'async_trait>>
    where
        'a: 'async_trait,
        Self: 'async_trait;
    fn begin_consensus_epoch<'a, 'life0, 'async_trait>(
        &'a self,
        dbtx: &'life0 mut DatabaseTransaction<'a>,
        consensus_items: Vec<(PeerId, Self::ConsensusItem)>,
        rng: impl 'async_trait + RngCore + CryptoRng + 'a
    ) -> Pin<Box<dyn Future<Output = ()> + 'async_trait>>
    where
        'a: 'async_trait,
        'life0: 'async_trait,
        Self: 'async_trait;
    fn build_verification_cache<'a>(
        &'a self,
        inputs: impl Iterator<Item = &'a Self::TxInput> + Send
    ) -> Self::VerificationCache;
    fn validate_input<'a>(
        &self,
        interconnect: &dyn ModuleInterconect,
        verification_cache: &Self::VerificationCache,
        input: &'a Self::TxInput
    ) -> Result<InputMeta<'a>, Self::Error>;
    fn apply_input<'a, 'b, 'c>(
        &'a self,
        interconnect: &'a dyn ModuleInterconect,
        dbtx: &mut DatabaseTransaction<'c>,
        input: &'b Self::TxInput,
        verification_cache: &Self::VerificationCache
    ) -> Result<InputMeta<'b>, Self::Error>;
    fn validate_output(
        &self,
        output: &Self::TxOutput
    ) -> Result<TransactionItemAmount, Self::Error>;
    fn apply_output<'a, 'b>(
        &'a self,
        dbtx: &mut DatabaseTransaction<'b>,
        output: &'a Self::TxOutput,
        out_point: OutPoint
    ) -> Result<TransactionItemAmount, Self::Error>;
    fn end_consensus_epoch<'a, 'life0, 'life1, 'async_trait>(
        &'a self,
        consensus_peers: &'life0 HashSet<PeerId>,
        dbtx: &'life1 mut DatabaseTransaction<'a>,
        rng: impl 'async_trait + RngCore + CryptoRng + 'a
    ) -> Pin<Box<dyn Future<Output = Vec<PeerId>> + 'async_trait>>
    where
        'a: 'async_trait,
        'life0: 'async_trait,
        'life1: 'async_trait,
        Self: 'async_trait;
    fn output_status(&self, out_point: OutPoint) -> Option<Self::TxOutputOutcome>;
    fn audit(&self, audit: &mut Audit);
    fn api_base_name(&self) -> &'static str;
    fn api_endpoints(&self) -> &'static [ApiEndpoint<Self>];

}

Required Associated Types

type Error

type TxInput: Send + Sync

type TxOutput

type TxOutputOutcome

type ConsensusItem

type VerificationCache

Required Methods

fn await_consensus_proposal<'a, 'async_trait>(
&'a self,
rng: impl 'async_trait + RngCore + CryptoRng + 'a
) -> Pin<Box<dyn Future<Output = ()> + 'async_trait>>where
'a: 'async_trait,
Self: 'async_trait,

Blocks until a new consensus_proposal is available.

fn consensus_proposal<'a, 'async_trait>(
&'a self,
rng: impl 'async_trait + RngCore + CryptoRng + 'a
) -> Pin<Box<dyn Future<Output = Vec<Self::ConsensusItem>> + 'async_trait>>where
'a: 'async_trait,
Self: 'async_trait,

This module’s contribution to the next consensus proposal

fn begin_consensus_epoch<'a, 'life0, 'async_trait>(
&'a self,
dbtx: &'life0 mut DatabaseTransaction<'a>,
consensus_items: Vec<(PeerId, Self::ConsensusItem)>,
rng: impl 'async_trait + RngCore + CryptoRng + 'a
) -> Pin<Box<dyn Future<Output = ()> + 'async_trait>>where
'a: 'async_trait,
'life0: 'async_trait,
Self: 'async_trait,

This function is called once before transaction processing starts. All module consensus items of this round are supplied as consensus_items. The batch will be committed to the database after all other modules ran begin_consensus_epoch, so the results are available when processing transactions.

fn build_verification_cache<'a>(
&'a self,
inputs: impl Iterator<Item = &'a Self::TxInput> + Send
) -> Self::VerificationCache

Some modules may have slow to verify inputs that would block transaction processing. If the slow part of verification can be modeled as a pure function not involving any system state we can build a lookup table in a hyper-parallelized manner. This function is meant for constructing such lookup tables.

fn validate_input<'a>(
&self,
interconnect: &dyn ModuleInterconect,
verification_cache: &Self::VerificationCache,
input: &'a Self::TxInput
) -> Result<InputMeta<'a>, Self::Error>

Validate a transaction input before submitting it to the unconfirmed transaction pool. This function has no side effects and may be called at any time. False positives due to outdated database state are ok since they get filtered out after consensus has been reached on them and merely generate a warning.

fn apply_input<'a, 'b, 'c>(
&'a self,
interconnect: &'a dyn ModuleInterconect,
dbtx: &mut DatabaseTransaction<'c>,
input: &'b Self::TxInput,
verification_cache: &Self::VerificationCache
) -> Result<InputMeta<'b>, Self::Error>

Try to spend a transaction input. On success all necessary updates will be part of the database batch. On failure (e.g. double spend) the batch is reset and the operation will take no effect.

This function may only be called after begin_consensus_epoch and before end_consensus_epoch. Data is only written to the database once all transaction have been processed.

fn validate_output(
&self,
output: &Self::TxOutput
) -> Result<TransactionItemAmount, Self::Error>

Validate a transaction output before submitting it to the unconfirmed transaction pool. This function has no side effects and may be called at any time. False positives due to outdated database state are ok since they get filtered out after consensus has been reached on them and merely generate a warning.

fn apply_output<'a, 'b>(
&'a self,
dbtx: &mut DatabaseTransaction<'b>,
output: &'a Self::TxOutput,
out_point: OutPoint
) -> Result<TransactionItemAmount, Self::Error>

Try to create an output (e.g. issue coins, peg-out BTC, …). On success all necessary updates to the database will be part of the batch. On failure (e.g. double spend) the batch is reset and the operation will take no effect.

The supplied out_point identifies the operation (e.g. a peg-out or coin issuance) and can be used to retrieve its outcome later using output_status.

This function may only be called after begin_consensus_epoch and before end_consensus_epoch. Data is only written to the database once all transactions have been processed.

fn end_consensus_epoch<'a, 'life0, 'life1, 'async_trait>(
&'a self,
consensus_peers: &'life0 HashSet<PeerId>,
dbtx: &'life1 mut DatabaseTransaction<'a>,
rng: impl 'async_trait + RngCore + CryptoRng + 'a
) -> Pin<Box<dyn Future<Output = Vec<PeerId>> + 'async_trait>>where
'a: 'async_trait,
'life0: 'async_trait,
'life1: 'async_trait,
Self: 'async_trait,

This function is called once all transactions have been processed and changes were written to the database. This allows running finalization code before the next epoch.

Passes in the consensus_peers that contributed to this epoch and returns a list of peers to drop if any are misbehaving.

fn output_status(&self, out_point: OutPoint) -> Option<Self::TxOutputOutcome>

Retrieve the current status of the output. Depending on the module this might contain data needed by the client to access funds or give an estimate of when funds will be available. Returns None if the output is unknown, NOT if it is just not ready yet.

fn audit(&self, audit: &mut Audit)

Queries the database and returns all assets and liabilities of the module.

Summing over all modules, if liabilities > assets then an error has occurred in the database and consensus should halt.

fn api_base_name(&self) -> &'static str

Defines the prefix for API endpoints defined by the module.

E.g. if the module’s base path is foo and it defines API endpoints bar and baz then these endpoints will be reachable under /foo/bar and /foo/baz.

fn api_endpoints(&self) -> &'static [ApiEndpoint<Self>]

Returns a list of custom API endpoints defined by the module. These are made available both to users as well as to other modules. They thus should be deterministic, only dependant on their input and the current epoch.