Trait fedimint_api::module::IModuleGen

pub trait IModuleGen: Debug {
    fn decoder(&self) -> DynDecoder;
    fn module_kind(&self) -> ModuleKind;
    fn init<'life0, 'life1, 'life2, 'async_trait>(
        &'life0 self,
        cfg: ServerModuleConfig,
        db: Database,
        env: &'life1 BTreeMap<OsString, OsString>,
        task_group: &'life2 mut TaskGroup
    ) -> Pin<Box<dyn Future<Output = Result<DynServerModule>> + Send + 'async_trait>>
    where
        Self: 'async_trait,
        'life0: 'async_trait,
        'life1: 'async_trait,
        'life2: 'async_trait;
    fn trusted_dealer_gen(
        &self,
        peers: &[PeerId],
        params: &ConfigGenParams
    ) -> BTreeMap<PeerId, ServerModuleConfig>;
    fn distributed_gen<'life0, 'life1, 'life2, 'life3, 'life4, 'life5, 'async_trait>(
        &'life0 self,
        connections: &'life1 MuxPeerConnections<ModuleInstanceId, DkgPeerMsg>,
        our_id: &'life2 PeerId,
        module_id: ModuleInstanceId,
        peers: &'life3 [PeerId],
        params: &'life4 ConfigGenParams,
        task_group: &'life5 mut TaskGroup
    ) -> Pin<Box<dyn Future<Output = Result<Cancellable<ServerModuleConfig>>> + Send + 'async_trait>>
    where
        Self: 'async_trait,
        'life0: 'async_trait,
        'life1: 'async_trait,
        'life2: 'async_trait,
        'life3: 'async_trait,
        'life4: 'async_trait,
        'life5: 'async_trait;
    fn to_config_response(&self, config: Value) -> Result<ModuleConfigResponse>;
    fn hash_client_module(&self, config: Value) -> Result<Hash>;
    fn validate_config(
        &self,
        identity: &PeerId,
        config: ServerModuleConfig
    ) -> Result<()>;
}

Interface for Module Generation

This trait contains the methods responsible for the module’s

initialization

config generation

config validation

Once the module configuration is ready, the module can be instantiated via [Self::init].

Required Methods

fn decoder(&self) -> DynDecoder

fn module_kind(&self) -> ModuleKind

fn init<'life0, 'life1, 'life2, 'async_trait>(
&'life0 self,
cfg: ServerModuleConfig,
db: Database,
env: &'life1 BTreeMap<OsString, OsString>,
task_group: &'life2 mut TaskGroup
) -> Pin<Box<dyn Future<Output = Result<DynServerModule>> + Send + 'async_trait>>where
Self: 'async_trait,
'life0: 'async_trait,
'life1: 'async_trait,
'life2: 'async_trait,

Initialize the DynServerModule instance from its config

fn trusted_dealer_gen(
&self,
peers: &[PeerId],
params: &ConfigGenParams
) -> BTreeMap<PeerId, ServerModuleConfig>

fn distributed_gen<'life0, 'life1, 'life2, 'life3, 'life4, 'life5, 'async_trait>(
&'life0 self,
connections: &'life1 MuxPeerConnections<ModuleInstanceId, DkgPeerMsg>,
our_id: &'life2 PeerId,
module_id: ModuleInstanceId,
peers: &'life3 [PeerId],
params: &'life4 ConfigGenParams,
task_group: &'life5 mut TaskGroup
) -> Pin<Box<dyn Future<Output = Result<Cancellable<ServerModuleConfig>>> + Send + 'async_trait>>where
Self: 'async_trait,
'life0: 'async_trait,
'life1: 'async_trait,
'life2: 'async_trait,
'life3: 'async_trait,
'life4: 'async_trait,
'life5: 'async_trait,

fn to_config_response(&self, config: Value) -> Result<ModuleConfigResponse>

fn hash_client_module(&self, config: Value) -> Result<Hash>

fn validate_config(
&self,
identity: &PeerId,
config: ServerModuleConfig
) -> Result<()>