Trait fedimint_api::module::ServerModule

pub trait ServerModule: Debug + Sized {
    type Gen: ModuleGen;
    type Decoder: Decoder;
    type VerificationCache: VerificationCache;

    // Required methods
    fn decoder(&self) -> Self::Decoder;
    fn versions(&self) -> (ModuleConsensusVersion, &[ApiVersion]);
    fn await_consensus_proposal<'a, 'life0, 'life1, 'async_trait>(
        &'a self,
        dbtx: &'life0 mut DatabaseTransaction<'life1>
    ) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>>
       where Self: 'async_trait,
             'a: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait;
    fn consensus_proposal<'a, 'life0, 'life1, 'async_trait>(
        &'a self,
        dbtx: &'life0 mut DatabaseTransaction<'life1>
    ) -> Pin<Box<dyn Future<Output = ConsensusProposal<<Self::Decoder as Decoder>::ConsensusItem>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'a: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait;
    fn begin_consensus_epoch<'a, 'b, 'life0, 'async_trait>(
        &'a self,
        dbtx: &'life0 mut DatabaseTransaction<'b>,
        consensus_items: Vec<(PeerId, <Self::Decoder as Decoder>::ConsensusItem)>
    ) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>>
       where Self: 'async_trait,
             'a: 'async_trait,
             'b: 'async_trait,
             'life0: 'async_trait;
    fn build_verification_cache<'a>(
        &'a self,
        inputs: impl Iterator<Item = &'a <Self::Decoder as Decoder>::Input> + Send
    ) -> Self::VerificationCache;
    fn validate_input<'a, 'b, 'life0, 'life1, 'life2, 'life3, 'async_trait>(
        &'life0 self,
        interconnect: &'life1 dyn ModuleInterconect,
        dbtx: &'life2 mut DatabaseTransaction<'b>,
        verification_cache: &'life3 Self::VerificationCache,
        input: &'a <Self::Decoder as Decoder>::Input
    ) -> Pin<Box<dyn Future<Output = Result<InputMeta, ModuleError>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'a: 'async_trait,
             'b: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait,
             'life2: 'async_trait,
             'life3: 'async_trait;
    fn apply_input<'a, 'b, 'c, 'life0, 'life1, 'async_trait>(
        &'a self,
        interconnect: &'a dyn ModuleInterconect,
        dbtx: &'life0 mut DatabaseTransaction<'c>,
        input: &'b <Self::Decoder as Decoder>::Input,
        verification_cache: &'life1 Self::VerificationCache
    ) -> Pin<Box<dyn Future<Output = Result<InputMeta, ModuleError>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'a: 'async_trait,
             'b: 'async_trait,
             'c: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait;
    fn validate_output<'life0, 'life1, 'life2, 'async_trait>(
        &'life0 self,
        dbtx: &'life1 mut DatabaseTransaction<'_>,
        output: &'life2 <Self::Decoder as Decoder>::Output
    ) -> Pin<Box<dyn Future<Output = Result<TransactionItemAmount, ModuleError>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait,
             'life2: 'async_trait;
    fn apply_output<'a, 'b, 'life0, 'async_trait>(
        &'a self,
        dbtx: &'life0 mut DatabaseTransaction<'b>,
        output: &'a <Self::Decoder as Decoder>::Output,
        out_point: OutPoint
    ) -> Pin<Box<dyn Future<Output = Result<TransactionItemAmount, ModuleError>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'a: 'async_trait,
             'b: 'async_trait,
             'life0: 'async_trait;
    fn end_consensus_epoch<'a, 'b, 'life0, 'life1, 'async_trait>(
        &'a self,
        consensus_peers: &'life0 HashSet<PeerId>,
        dbtx: &'life1 mut DatabaseTransaction<'b>
    ) -> Pin<Box<dyn Future<Output = Vec<PeerId>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'a: 'async_trait,
             'b: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait;
    fn output_status<'life0, 'life1, 'life2, 'async_trait>(
        &'life0 self,
        dbtx: &'life1 mut DatabaseTransaction<'life2>,
        out_point: OutPoint
    ) -> Pin<Box<dyn Future<Output = Option<<Self::Decoder as Decoder>::OutputOutcome>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait,
             'life2: 'async_trait;
    fn audit<'life0, 'life1, 'life2, 'life3, 'async_trait>(
        &'life0 self,
        dbtx: &'life1 mut DatabaseTransaction<'life2>,
        audit: &'life3 mut Audit
    ) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait,
             'life2: 'async_trait,
             'life3: 'async_trait;
    fn api_endpoints(&self) -> Vec<ApiEndpoint<Self>>;

    // Provided method
    fn module_kind() -> ModuleKind { ... }

}

Required Associated Types

type Gen: ModuleGen

type Decoder: Decoder

type VerificationCache: VerificationCache

Required Methods

fn decoder(&self) -> Self::Decoder

fn versions(&self) -> (ModuleConsensusVersion, &[ApiVersion])

Module consensus version this module is running with and the API versions it supports in it

fn await_consensus_proposal<'a, 'life0, 'life1, 'async_trait>( &'a self, dbtx: &'life0 mut DatabaseTransaction<'life1> ) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>>where Self: 'async_trait, 'a: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

Blocks until a new consensus_proposal is available.

fn consensus_proposal<'a, 'life0, 'life1, 'async_trait>( &'a self, dbtx: &'life0 mut DatabaseTransaction<'life1> ) -> Pin<Box<dyn Future<Output = ConsensusProposal<<Self::Decoder as Decoder>::ConsensusItem>> + Send + 'async_trait>>where Self: 'async_trait, 'a: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

This module’s contribution to the next consensus proposal

fn begin_consensus_epoch<'a, 'b, 'life0, 'async_trait>( &'a self, dbtx: &'life0 mut DatabaseTransaction<'b>, consensus_items: Vec<(PeerId, <Self::Decoder as Decoder>::ConsensusItem)> ) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>>where Self: 'async_trait, 'a: 'async_trait, 'b: 'async_trait, 'life0: 'async_trait,

This function is called once before transaction processing starts. All module consensus items of this round are supplied as consensus_items. The database transaction will be committed to the database after all other modules ran begin_consensus_epoch, so the results are available when processing transactions.

fn build_verification_cache<'a>( &'a self, inputs: impl Iterator<Item = &'a <Self::Decoder as Decoder>::Input> + Send ) -> Self::VerificationCache

Some modules may have slow to verify inputs that would block transaction processing. If the slow part of verification can be modeled as a pure function not involving any system state we can build a lookup table in a hyper-parallelized manner. This function is meant for constructing such lookup tables.

fn validate_input<'a, 'b, 'life0, 'life1, 'life2, 'life3, 'async_trait>( &'life0 self, interconnect: &'life1 dyn ModuleInterconect, dbtx: &'life2 mut DatabaseTransaction<'b>, verification_cache: &'life3 Self::VerificationCache, input: &'a <Self::Decoder as Decoder>::Input ) -> Pin<Box<dyn Future<Output = Result<InputMeta, ModuleError>> + Send + 'async_trait>>where Self: 'async_trait, 'a: 'async_trait, 'b: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait, 'life2: 'async_trait, 'life3: 'async_trait,

Validate a transaction input before submitting it to the unconfirmed transaction pool. This function has no side effects and may be called at any time. False positives due to outdated database state are ok since they get filtered out after consensus has been reached on them and merely generate a warning.

fn apply_input<'a, 'b, 'c, 'life0, 'life1, 'async_trait>( &'a self, interconnect: &'a dyn ModuleInterconect, dbtx: &'life0 mut DatabaseTransaction<'c>, input: &'b <Self::Decoder as Decoder>::Input, verification_cache: &'life1 Self::VerificationCache ) -> Pin<Box<dyn Future<Output = Result<InputMeta, ModuleError>> + Send + 'async_trait>>where Self: 'async_trait, 'a: 'async_trait, 'b: 'async_trait, 'c: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

Try to spend a transaction input. On success all necessary updates will be part of the database transaction. On failure (e.g. double spend) the database transaction is rolled back and the operation will take no effect.

This function may only be called after begin_consensus_epoch and before end_consensus_epoch. Data is only written to the database once all transactions have been processed.

fn validate_output<'life0, 'life1, 'life2, 'async_trait>( &'life0 self, dbtx: &'life1 mut DatabaseTransaction<'_>, output: &'life2 <Self::Decoder as Decoder>::Output ) -> Pin<Box<dyn Future<Output = Result<TransactionItemAmount, ModuleError>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait, 'life2: 'async_trait,

Validate a transaction output before submitting it to the unconfirmed transaction pool. This function has no side effects and may be called at any time. False positives due to outdated database state are ok since they get filtered out after consensus has been reached on them and merely generate a warning.

fn apply_output<'a, 'b, 'life0, 'async_trait>( &'a self, dbtx: &'life0 mut DatabaseTransaction<'b>, output: &'a <Self::Decoder as Decoder>::Output, out_point: OutPoint ) -> Pin<Box<dyn Future<Output = Result<TransactionItemAmount, ModuleError>> + Send + 'async_trait>>where Self: 'async_trait, 'a: 'async_trait, 'b: 'async_trait, 'life0: 'async_trait,

Try to create an output (e.g. issue notes, peg-out BTC, …). On success all necessary updates to the database will be part of the database transaction. On failure (e.g. double spend) the database transaction is rolled back and the operation will take no effect.

The supplied out_point identifies the operation (e.g. a peg-out or note issuance) and can be used to retrieve its outcome later using output_status.

This function may only be called after begin_consensus_epoch and before end_consensus_epoch. Data is only written to the database once all transactions have been processed.

fn end_consensus_epoch<'a, 'b, 'life0, 'life1, 'async_trait>( &'a self, consensus_peers: &'life0 HashSet<PeerId>, dbtx: &'life1 mut DatabaseTransaction<'b> ) -> Pin<Box<dyn Future<Output = Vec<PeerId>> + Send + 'async_trait>>where Self: 'async_trait, 'a: 'async_trait, 'b: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

This function is called once all transactions have been processed and changes were written to the database. This allows running finalization code before the next epoch.

Passes in the consensus_peers that contributed to this epoch and returns a list of peers to drop if any are misbehaving.

fn output_status<'life0, 'life1, 'life2, 'async_trait>( &'life0 self, dbtx: &'life1 mut DatabaseTransaction<'life2>, out_point: OutPoint ) -> Pin<Box<dyn Future<Output = Option<<Self::Decoder as Decoder>::OutputOutcome>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait, 'life2: 'async_trait,

Retrieve the current status of the output. Depending on the module this might contain data needed by the client to access funds or give an estimate of when funds will be available. Returns None if the output is unknown, NOT if it is just not ready yet.

fn audit<'life0, 'life1, 'life2, 'life3, 'async_trait>( &'life0 self, dbtx: &'life1 mut DatabaseTransaction<'life2>, audit: &'life3 mut Audit ) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait, 'life2: 'async_trait, 'life3: 'async_trait,

Queries the database and returns all assets and liabilities of the module.

Summing over all modules, if liabilities > assets then an error has occurred in the database and consensus should halt.

fn api_endpoints(&self) -> Vec<ApiEndpoint<Self>>

Returns a list of custom API endpoints defined by the module. These are made available both to users as well as to other modules. They thus should be deterministic, only dependant on their input and the current epoch.

Provided Methods

fn module_kind() -> ModuleKind