Module fedimint_api::task::imp

Structs

RwLock

An asynchronous reader-writer lock.

RwLockReadGuard

RAII structure used to release the shared read access of a lock when dropped.

RwLockWriteGuard

RAII structure used to release the exclusive write access of a lock when dropped.

Functions

block_in_place

sleep

sleep_until

spawn 🔒

spawn_local 🔒

timeout