Trait fedimint_api::BitcoinHash

pub trait BitcoinHash: Copy + Clone + PartialEq<Self> + Eq + PartialOrd<Self> + Ord + Hash + Debug + Display + LowerHex + Index<RangeFull, Output = [u8], Output = [u8], Output = [u8], Output = [u8], Output = u8> + Index<RangeFrom<usize>> + Index<RangeTo<usize>> + Index<Range<usize>> + Index<usize> + Borrow<[u8]> {
    type Engine: HashEngine;
    type Inner: FromHex;

    const LEN: usize;
    const DISPLAY_BACKWARD: bool = false;

    fn from_engine(e: Self::Engine) -> Self;
    fn from_slice(sl: &[u8]) -> Result<Self, Error>;
    fn into_inner(self) -> Self::Inner;
    fn as_inner(&self) -> &Self::Inner;
    fn from_inner(inner: Self::Inner) -> Self;
    fn all_zeros() -> Self;

    fn engine() -> Self::Engine { ... }
    fn hash(data: &[u8]) -> Self { ... }
}

Trait which applies to hashes of all types.

Required Associated Types

type Engine: HashEngine

A hashing engine which bytes can be serialized into. It is expected to implement the io::Write trait, and to never return errors under any conditions.

type Inner: FromHex

The byte array that represents the hash internally.

Required Associated Constants

const LEN: usize

Length of the hash, in bytes.

Provided Associated Constants

const DISPLAY_BACKWARD: bool = false

Flag indicating whether user-visible serializations of this hash should be backward. For some reason Satoshi decided this should be true for Sha256dHash, so here we are.

Required Methods

fn from_engine(e: Self::Engine) -> Self

Produces a hash from the current state of a given engine.

fn from_slice(sl: &[u8]) -> Result<Self, Error>

Copies a byte slice into a hash object.

fn into_inner(self) -> Self::Inner

Unwraps the hash and returns the underlying byte array.

fn as_inner(&self) -> &Self::Inner

Unwraps the hash and returns a reference to the underlying byte array.

fn from_inner(inner: Self::Inner) -> Self

Constructs a hash from the underlying byte array.

fn all_zeros() -> Self

Returns an all zero hash.

An all zeros hash is a made up construct because there is not a known input that can create it, however it is used in various places in Bitcoin e.g., the Bitcoin genesis block’s previous blockhash and the coinbase transaction’s outpoint txid.

Provided Methods

fn engine() -> Self::Engine

Constructs a new engine.

fn hash(data: &[u8]) -> Self

Hashes some bytes.