Trait fedimint_api::NumPeers

pub trait NumPeers {
    fn total(&self) -> usize;

    fn max_evil(&self) -> usize { ... }
    fn one_honest(&self) -> usize { ... }
    fn degree(&self) -> usize { ... }
    fn threshold(&self) -> usize { ... }
}

for consensus-related calculations given the number of peers

Required Methods

fn total(&self) -> usize

Provided Methods

fn max_evil(&self) -> usize

number of peers that can be evil without disrupting the federation

fn one_honest(&self) -> usize

number of peers to select such that one is honest (under our assumptions)

fn degree(&self) -> usize

Degree of a underlying polynomial to require threshold signatures

fn threshold(&self) -> usize

number of peers required for a signature