Crate fedimint_api

Re-exports

pub use module::FederationModule;

pub use module::InputMeta;

Modules

config

db

encoding

This module defines a binary encoding interface which is more suitable for consensus critical encoding thant e.g. bincode. Over time all structs that need to be encoded to binary will be migrated to this interface.

macros

module

rand

task

Macros

__api_endpoint

Example

dyn_newtype_define

Define “dyn newtype” (a newtype over dyn Trait)

dyn_newtype_impl_dyn_clone_passhthrough

Implement Clone on a “dyn newtype”

Structs

Amount

Represents an amount of BTC inside the system. The base denomination is milli satoshi for now, this is also why the amount type from rust-bitcoin isn’t used instead.

OutPoint

PeerId

TransactionId

A transaction id for peg-ins, peg-outs and reissuances

Enums

ParseAmountError

Traits

BitcoinHash

Trait which applies to hashes of all types

NumPeers

for consensus-related calculations given the number of peers