Trait fedimint_api::config::TypedServerModuleConfig

pub trait TypedServerModuleConfig: DeserializeOwned + Serialize {
    type Local: DeserializeOwned + Serialize;
    type Private: DeserializeOwned + Serialize;
    type Consensus: TypedServerModuleConsensusConfig;

    fn from_parts(
        local: Self::Local,
        private: Self::Private,
        consensus: Self::Consensus
    ) -> Self;
    fn to_parts(
        self
    ) -> (ModuleKind, Self::Local, Self::Private, Self::Consensus);
    fn validate_config(&self, identity: &PeerId) -> Result<()>;

    fn to_erased(self) -> ServerModuleConfig { ... }
}

Module (server side) config

Required Associated Types

type Local: DeserializeOwned + Serialize

Local non-consensus, not security-sensitive settings

type Private: DeserializeOwned + Serialize

Private for this federation member data that are security sensitive and will be encrypted at rest

type Consensus: TypedServerModuleConsensusConfig

Shared consensus-critical config

Required Methods

fn from_parts(
local: Self::Local,
private: Self::Private,
consensus: Self::Consensus
) -> Self

Assemble from the three functionally distinct parts

fn to_parts(self) -> (ModuleKind, Self::Local, Self::Private, Self::Consensus)

Split the config into its three functionally distinct parts

fn validate_config(&self, identity: &PeerId) -> Result<()>

Validate the config

Provided Methods

fn to_erased(self) -> ServerModuleConfig

Turn the typed config into type-erased version