Module fedimint_api::core

Fedimint Core API (common) module interface

Fedimint supports externally implemented modules.

This (Rust) module defines common interoperability types and functionality that is used on both client and sever side.

Modules

client

encode

server

Fedimint Core Server module interface

Macros

erased_eq_no_instance_id 🔒

newtype_impl_display_passthrough 🔒

Implementes the Display trait for dyn newtypes whose traits implement Display

newtype_impl_display_passthrough_with_instance_id 🔒

newtype_impl_eq_passthrough_with_instance_id 🔒

try 🔒

try 🔒

Structs

DynDecoder

Decoder for module associated types

DynInput

An owned, immutable input to a Transaction

DynModuleConsensusItem

An owned, immutable output of a Transaction before it was finalized

DynOutput

An owned, immutable output of a Transaction

DynOutputOutcome

An owned, immutable output of a Transaction before it was finalized

KeyPair

Opaque data structure that holds a keypair consisting of a secret and a public key.

ModuleKind

A type of a module

Signature

Transaction

Transaction that was already signed

__Visitor 🔒

Enums

FinalizationError

Constants

LEGACY_HARDCODED_INSTANCE_ID_LN

LEGACY_HARDCODED_INSTANCE_ID_MINT

LEGACY_HARDCODED_INSTANCE_ID_WALLET

MODULE_INSTANCE_ID_GLOBAL

Special ID we use for global dkg

Traits

Decoder

Module Decoder trait

IDecoder

IInput

Something that can be an DynInput in a Transaction

IModuleConsensusItem

IOutput

Something that can be an DynOutput in a Transaction

IOutputOutcome

Input

ModuleConsensusItem

Output

OutputOutcome

Type Definitions

ModuleInstanceId

Module instance ID