Module fedimint_api::core

Fedimint Core API (common) module interface

Fedimint supports externally implemented modules.

This (Rust) module defines common interoperability types and functionality that is used on both client and sever side.

Modules

client

encode

server

Fedimint Core Server module interface

Macros

erased_eq 🔒

newtype_impl_display_passthrough 🔒

Implementes the Display trait for dyn newtypes whose traits implement Display

newtype_impl_eq_passthrough 🔒

Structs

ConsensusItem

An owned, immutable output of a Transaction before it was finalized

Decoder

Decoder for module associated types

Input

An owned, immutable input to a Transaction

KeyPair

Opaque data structure that holds a keypair consisting of a secret and a public key.

Output

An owned, immutable output of a Transaction

OutputOutcome

An owned, immutable output of a Transaction before it was finalized

Signature

Transaction

Transaction that was already signed

Enums

FinalizationError

Constants

MODULE_KEY_GLOBAL

MODULE_KEY_LN

MODULE_KEY_MINT

MODULE_KEY_WALLET

Temporary constant for the modules we already have

Traits

ModuleConsensusItem

ModuleDecode

ModuleInput

Something that can be an Input in a Transaction

ModuleOutput

Something that can be an Output in a Transaction

ModuleOutputOutcome

PluginConsensusItem

PluginDecode

Module Decoder trait

PluginInput

PluginOutput

PluginOutputOutcome

Type Definitions

ModuleKey

A module key identifing a module