Trait fedimint_api::core::ModuleDecode

pub trait ModuleDecode {
    fn decode_spendable_output(
        &self,
        r: &mut dyn Read
    ) -> Result<SpendableOutput, DecodeError>;
    fn decode_input(&self, r: &mut dyn Read) -> Result<Input, DecodeError>;
    fn decode_output(&self, r: &mut dyn Read) -> Result<Output, DecodeError>;
    fn decode_pending_output(
        &self,
        r: &mut dyn Read
    ) -> Result<PendingOutput, DecodeError>;
    fn decode_output_outcome(
        &self,
        r: &mut dyn Read
    ) -> Result<OutputOutcome, DecodeError>;
    fn decode_consensus_item(
        &self,
        r: &mut dyn Read
    ) -> Result<ConsensusItem, DecodeError>;
}

Required Methods

fn decode_spendable_output(
&self,
r: &mut dyn Read
) -> Result<SpendableOutput, DecodeError>

Decode SpendableOutput compatible with this module, after the module key prefix was already decoded

fn decode_input(&self, r: &mut dyn Read) -> Result<Input, DecodeError>

Decode Input compatible with this module, after the module key prefix was already decoded

fn decode_output(&self, r: &mut dyn Read) -> Result<Output, DecodeError>

Decode Output compatible with this module, after the module key prefix was already decoded

fn decode_pending_output(
&self,
r: &mut dyn Read
) -> Result<PendingOutput, DecodeError>

Decode PendingOutput compatible with this module, after the module key prefix was already decoded

fn decode_output_outcome(
&self,
r: &mut dyn Read
) -> Result<OutputOutcome, DecodeError>

Decode OutputOutcome compatible with this module, after the module key prefix was already decoded

fn decode_consensus_item(
&self,
r: &mut dyn Read
) -> Result<ConsensusItem, DecodeError>

Decode ConsensusItem compatible with this module, after the module key prefix was already decoded