Module fedimint_api::db

Modules

mem_impl

tests 🔒

Structs

Database

A handle to a type-erased database implementation

DatabaseDeleteOperation

DatabaseInsertOperation

DatabaseTransaction

A handle to a type-erased database implementation

Enums

DatabaseOperation

DecodingError

TestDbKeyPrefix

Traits

DatabaseKey

DatabaseKeyPrefix

DatabaseKeyPrefixConst

DatabaseValue

IDatabase

IDatabaseTransaction

Fedimint requires that the database implementation implement Snapshot Isolation. Snapshot Isolation is a database isolation level that guarantees consistent reads from the time that the snapshot was created (at transaction creation time). Transactions with Snapshot Isolation level will only commit if there has been no write to the modified keys since the snapshot (i.e. write-write conflicts are prevented).

SerializableDatabaseValue

Functions

expect_write_conflict

verify_commit

verify_find_by_prefix

verify_insert_elements

verify_phantom_entry

verify_prevent_dirty_reads

verify_prevent_nonrepeatable_reads

verify_read_own_writes

verify_remove_existing

verify_remove_nonexisting

verify_rollback_to_savepoint

Type Definitions

PrefixIter