Trait fedimint_api::db::IDatabaseTransaction

pub trait IDatabaseTransaction<'a>: 'a + Send {
    // Required methods
    fn raw_insert_bytes<'life0, 'life1, 'async_trait>(
        &'life0 mut self,
        key: &'life1 [u8],
        value: Vec<u8>
    ) -> Pin<Box<dyn Future<Output = Result<Option<Vec<u8>>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait;
    fn raw_get_bytes<'life0, 'life1, 'async_trait>(
        &'life0 mut self,
        key: &'life1 [u8]
    ) -> Pin<Box<dyn Future<Output = Result<Option<Vec<u8>>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait;
    fn raw_remove_entry<'life0, 'life1, 'async_trait>(
        &'life0 mut self,
        key: &'life1 [u8]
    ) -> Pin<Box<dyn Future<Output = Result<Option<Vec<u8>>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait;
    fn raw_find_by_prefix<'life0, 'life1, 'async_trait>(
        &'life0 mut self,
        key_prefix: &'life1 [u8]
    ) -> Pin<Box<dyn Future<Output = PrefixStream<'_>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait;
    fn commit_tx<'async_trait>(
        self: Box<Self>
    ) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>
       where Self: 'async_trait;
    fn rollback_tx_to_savepoint<'life0, 'async_trait>(
        &'life0 mut self
    ) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn set_tx_savepoint<'life0, 'async_trait>(
        &'life0 mut self
    ) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;

    // Provided method
    fn raw_remove_by_prefix<'life0, 'life1, 'async_trait>(
        &'life0 mut self,
        key_prefix: &'life1 [u8]
    ) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait { ... }
}

Fedimint requires that the database implementation implement Snapshot Isolation. Snapshot Isolation is a database isolation level that guarantees consistent reads from the time that the snapshot was created (at transaction creation time). Transactions with Snapshot Isolation level will only commit if there has been no write to the modified keys since the snapshot (i.e. write-write conflicts are prevented).

Specifically, Fedimint expects the database implementation to prevent the following anamolies:

Non-Readable Write: TX1 writes (K1, V1) at time t but cannot read (K1, V1) at time (t + i)

Dirty Read: TX1 is able to read TX2’s uncommitted writes.

Non-Repeatable Read: TX1 reads (K1, V1) at time t and retrieves (K1, V2) at time (t + i) where V1 != V2.

Phantom Record: TX1 retrieves X number of records for a prefix at time t and retrieves Y number of records for the same prefix at time (t + i).

Lost Writes: TX1 writes (K1, V1) at the same time as TX2 writes (K1, V2). V2 overwrites V1 as the value for K1 (write-write conflict).

Type	Non-Readable Write	Dirty Read	Non-Repeatable Read	Phantom Record	Lost Writes

MemoryDB	Prevented	Prevented	Prevented	Prevented	Possible

RocksDB	Prevented	Prevented	Prevented	Prevented	Prevented

Sqlite	Prevented	Prevented	Prevented	Prevented	Prevented

Required Methods

fn raw_insert_bytes<'life0, 'life1, 'async_trait>( &'life0 mut self, key: &'life1 [u8], value: Vec<u8> ) -> Pin<Box<dyn Future<Output = Result<Option<Vec<u8>>>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

fn raw_get_bytes<'life0, 'life1, 'async_trait>( &'life0 mut self, key: &'life1 [u8] ) -> Pin<Box<dyn Future<Output = Result<Option<Vec<u8>>>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

fn raw_remove_entry<'life0, 'life1, 'async_trait>( &'life0 mut self, key: &'life1 [u8] ) -> Pin<Box<dyn Future<Output = Result<Option<Vec<u8>>>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

fn raw_find_by_prefix<'life0, 'life1, 'async_trait>( &'life0 mut self, key_prefix: &'life1 [u8] ) -> Pin<Box<dyn Future<Output = PrefixStream<'_>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

fn commit_tx<'async_trait>( self: Box<Self> ) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>where Self: 'async_trait,

fn rollback_tx_to_savepoint<'life0, 'async_trait>( &'life0 mut self ) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

fn set_tx_savepoint<'life0, 'async_trait>( &'life0 mut self ) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

Create a savepoint during the transaction that can be rolled back to using rollback_tx_to_savepoint. Rolling back to the savepoint will atomically remove the writes that were applied since the savepoint was created.

Warning: Avoid using this in fedimint client code as not all database transaction implementations will support setting a savepoint during a transaction.

Provided Methods

fn raw_remove_by_prefix<'life0, 'life1, 'async_trait>( &'life0 mut self, key_prefix: &'life1 [u8] ) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

Default implementation is a combination of Self::raw_find_by_prefix + loop over Self::raw_remove_entry